server/requirements.txt
openenv-core
pydantic
